Fix for https://bitbucket.org/wjiang/gpssim/issue/1/unhandled-exception-while-serving-an
Moved the gpssim.py file into the gpssim folder, the init in this file imports the the GpsSim class so existing imports of this package should not break. This needed to happen so that unit tests can import the sim. Created a constants file in the hopes that the many strings littered throughout this package can one day be constafied. Added a unit test for my fix.